Rename India-District-Elevation-Explorer.js to India-District-Elevation-Explorer.js
This application allows you to explore the elevation data of various districts in India. 

Select a state and a district to visualize the SRTM DEM data. You can see the minimum and maximum elevation values for the selected district, and click on the map to get the elevation at any specific point within the district.